@@ -1,5 +1,5 @@
 import React from "react";
-import ReactDOM from "react-dom";
-import App from "App";
+import Root from "root";
+import { createRoot } from 'react-dom/client';
 
-ReactDOM.render(<App />, document.getElementById("app"));
+createRoot(document.getElementById("app") as HTMLElement).render(<Root />);
